requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
flask==2.0.2
gunicorn==20.1.0
python-decouple==3.6
pyjwt==2.4.0
python-dotenv==0.20.0
#PyJWT==1.4.2
#Flask-login==0.6.2
#Flask-MySQLdb==1.0.1
#Flask-WTF==1.0.1